'run-modified-tests FILES' works like you'd want
It tracks down the right tester file for FILES and runs them.
As a bonus, when run in this fashion, it won't blow away the
cachefile of failed tests. This allows you to work through
test failures one-by-one, yet still eventually

  run-modified-tests --retry

to rerun the whole failed suite.

If you want to forcibly not overwrite the cachefile, you
may now specify

  run-modified-tests --nosave